Escapes markdown characters in clean_card_text to fix how Maxx's flavor text displays
Previously discord tried to interpret the asterixis in Maxx's flavor text as markdown. This fixes that. I'm not sure if there's any other cards that escaping markdown characters matters for.

The lines of text in clean_card_text to fix lists were also indented an extra time, causing them to run every time in the above for loop.